Various ViewModel wrangling.
At this point it's not working yet, as I feel I need to better clarify
how I'd use ViewModel in the activity and fragments in the app, which
uses the Single-Activity architecture.

I'm going to stop the ViewModel work for a bit, to step back and add
Navigation component to the app, which will support the Single-Activity
architecture better.