Test for sending data from SubjectFragment to SubjectsBottomDialogFragment.
This is using constructor parameter and does not seem ideal. Because
sometimes Android want to try to restore a fragment automatically, and
in that case it ignores parameter and uses no-parameter constructor
instead. Looks like it's better to set and get data in preferences
instead. Will work on that next.