In Subject Note Edit screen, add alert when user presses back/up.
The alert is to ask for saving changes (or not). This commit already includes the saving functionality as well, all good. The more crucial thing to note here is that Up and Back button have to be handled separately here (Back in SubjectNoteEditFragment.kt, and Up in the MainActivity.kt).