Wrangle MaterialDatePicker UTC day shenanigan.
This app uses MaterialDatePicker, which has a "gotcha" where if someone picks a date from it, the resulting Long value is actually the timestamp that equals to UTC+0 date at 00:00. So for example, if someone with Hawaii as system timezone (UTC-10) picks a December 25 on the picker, the returned value from the picker is actually December 25 at 00:00 UTC+0 time, *not* December 25 at 00:00 UTC-10 time.

This can cause issue because if the returned value is then displayed on the UI according to system's timezone (for example using SimpleDateFormat()), then the formatted string will actually say December 24, 14:00, which is not what the user picked in the first place.

This commit includes convertStartDayUTCMillisToStartDaySystemTimezoneMillis() to deal with that.